requirements.txt
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
click==8.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
htmltools==0.6.0
idna==3.11
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
linkify-it-py==2.0.3
markdown-it-py==4.0.0
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
mizani==0.14.3
narwhals==2.10.1
nest-asyncio==1.6.0
numpy==2.3.3
orjson==3.11.4
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.4.0
plotnine==0.15.1
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
questionary==2.1.1
scipy==1.16.3
setuptools==80.9.0
shiny==1.5.0
shinychat==0.2.8
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.5
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
